backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pyTigerGraph==1.7.4
pandas==2.2.3
numpy==2.2.4
pyshp==2.3.1
python-dotenv==1.0.1
websockets==14.2
